even fibonacci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	long long int j, k, l, sum;
+	long int j, k, l, sum;
 
 	j = 0;
 	k = 1;
@@ -21,6 +21,6 @@ int main(void)
 		if (l % 2 == 0)
 			sum += l;
 	}
-	printf("%lld\n", sum);
+	printf("%ld\n", sum);
 	return (0);
 }
